+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -128,7 +128,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kdx3cm3079j7gyex91wqhv8t"),
+  _id: z.string().default("conv_01kdyvszkw27tg5fs5zerxyywh"),
   entityId: z.string(),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
@@ -161,7 +161,7 @@ export const ConversationResponse$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ConversationResponse
 > = z.object({
-  id: z.string().default("conv_01kdx3cm3079j7gyex91wqhv8t"),
+  id: z.string().default("conv_01kdyvszkw27tg5fs5zerxyywh"),
   entityId: z.string(),
   kind: ConversationResponseKind$outboundSchema,
   displayName: z.string(),