+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -244,7 +244,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kdx3cm323m2rc9rmhf2qjc6y"),
+  _id: z.string().default("conv_01kdyvszkzp3et86kd0b1zm7rn"),
   entityId: z.string(),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
@@ -278,7 +278,7 @@ export const CreateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kdx3cm323m2rc9rmhf2qjc6y"),
+  id: z.string().default("conv_01kdyvszkzp3et86kd0b1zm7rn"),
   entityId: z.string(),
   kind: CreateConversationKind$outboundSchema,
   displayName: z.string(),