+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -307,7 +307,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kdx3cm36gm39b1bjm1xksczb"),
+  _id: z.string().default("conv_01kdyvszm5b3av07z0ym2wp537"),
   entityId: z.string(),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
@@ -341,7 +341,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kdx3cm36gm39b1bjm1xksczb"),
+  id: z.string().default("conv_01kdyvszm5b3av07z0ym2wp537"),
   entityId: z.string(),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),