+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -256,7 +256,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-01T15:38:46.209Z",
+    "2026-01-02T08:04:42.035Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -295,7 +295,7 @@ export const UpdateDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2026-01-01T15:38:46.209Z"))
+  updated: z.date().default(() => new Date("2026-01-02T08:04:42.035Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {